Minor
Ok so there was no problem.  It just happens that coincdently the ipad I was using for testing had two other devices using the same tag number.  This was throwing lost mode off, but get info is ready for that sort of problem...  I should probably look at adapting lostmode to realize when two devices are reported back and figure out how to deal with them.