Update lostmodefun.sh
Debugged my mass play sound.  It wasn't working... but is now.  Yes you really can send a handful of UDIDs to mosyle for that command and it will play sound on them all.

Also added code to tell you when you enable Lost mode sorta when the device last checked in, color coded, to give you an idea of how successful the command will be.