MacOS 12.3 Takes Python away
Right now in the 12.3 beta release python is gone.  I use python as a converted for epoch time so the alternative is I use Munki python.  Its popular.. so if you have Munki installed we use it.  If you don't then you have to modify config yourself and path out your own python blend.  In the future a better detect routine may be helpful.