Serial Search
Added ability to give a text file of serials and query the Mosyle database to see if the serial is listed.  If so it outputs to a file.  This is great when you have a file of devices you got rid of for buy back but need to make sure they are not tieing up licenses anymore.

When  a device is removed from ASM it does NOT remove from Mosyle.