JSON to CSV Fun
Working with a new method now for exporting the Mosyle Data to tab delimited file.  This was previously done some monsterous perl code, but the problem with that code its VERY finicky to special characters and when we decide to ask for another field of data the perl code also needs more fields... but it can't do more than 9 so now we must sacrifice data points.

Now we dump to file, prep the file so it looks like standard JSON, and then pass it to Python to output the same end result we had before, but this method doesn't care about special chars and doesn't need changed when we change what data we are dropping.

This new method is in place with the MACDump today, but will quickly be ported for iPads and User dumps when I can.