Added Bearer Token to Lost Mode
This is working, but its ugly...  I donT' think if a command is successful anymore that the API tells you it was...  Like I think you only get feedback when it bombs.  Will need to look at that closer, but as far as I can tell this stuff is working.