Update mosdatadump-IOS.sh
So I learned today that there is a relatively simple way to query say all iPads in one snap.  All Macs in a snap...  This is my VERY scratch way of getting a single export and dropping to file.  I tried to keep as much original as possible and only change what I had to...  I can aways make it better after I undersrtand it better.