Update mosbasic
Minor changes to correct how the configs and what not are found.  This was critical because I want to seperate certain configs and what not when I work with a Docker (or I think I do) and the way this was before it wasn't good.  There also was a bug in how we found BAGCLI_WORKDIR in that if MOSBasic is kept in a directory called MOSBASIC so effectively it was MOSBasic/MOSBasic then both would be stripped in the the pathing call.

The Other changes are accidents really...  My cut and pate dropped all the extra clutter characters.