Pt1 Class Support
This is the first of a few posts I'm sure where I am trying to create/build in support for READING class information from Mosyle.  I've recently stumbled into a need to see whos in what class and act on it from the API so here we go.

You will notice that the way I process JSON to CSV in this setup is very different vs how I do others.  When MOSBasic started my goal was to create a tool, strictly simple shell, that could bridge the command line with the webpage.  Now MUSKY is my focus.. and since MOSBasic is a huge corner stone of that system I feel it will be more important to embrace modern  so there may be a few times going forward that extra dependacies could be required.  For instance this addition uses jq to convert the JSON.