Upping the Version and adding the commands
I tried to document as well as I could in the code itself as I went along.  I also think if you type MOSBASIC with no arguement it explains fairly nice though I know a wiki page is in order.

HAY ME-> We need a safety valve still in getinfo-classes.sh so if dump data doesn't exist we make a dump happen!